Adds shedding of platform inertia and better platform collision.
When you are on a platform that is moving right and you jump and hold left,
you will accelerate to the left until the platform's angular velocity is gone.
The opposite goes for moving left.

Better platform collision is achieved by calculating the angle of the player
based on the sprite width and the distance from the center of the planet.
Half of the player sprite's angular width is used to pad platforms on
either side so that the entire sprite has to be off the platform before
you fall, instead of half of it. This makes jumping on and off platforms
less frustrating. Platforms behave more intuitively.